porject_SCENE_DETECTION/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.10.0=pypi_0
accelerate=0.23.0=pypi_0
aiofiles=23.1.0=pypi_0
aiogram=2.14=pypi_0
aiogram-media-group=0.5.1=pypi_0
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.0.1=pypi_0
annotated-types=0.5.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.0.0=pypi_0
apscheduler=3.6.3=pypi_0
asttokens=2.2.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
blinker=1.6.2=pypi_0
blis=0.7.10=pypi_0
braceexpand=0.1.7=pypi_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2023.05.30=haa95532_0
cachetools=4.2.2=pypi_0
catalogue=2.0.9=pypi_0
certifi=2023.7.22=py38haa95532_0
cffi=1.15.1=py38h2bbff1b_3
cfgv=3.4.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=py38haa95532_0
confection=0.1.1=pypi_0
contexttimer=0.3.3=pypi_0
contourpy=1.1.0=pypi_0
cryptography=41.0.2=py38hac1b9e3_0
cuda-cccl=12.2.128=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.2.128=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pypi_0
cymem=2.0.7=pypi_0
decorator=5.1.1=pypi_0
decord=0.6.0=pypi_0
diffusers=0.16.0=pypi_0
distlib=0.3.7=pypi_0
einops=0.6.1=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=1.2.0=pypi_0
fairscale=0.4.4=pypi_0
filelock=3.12.2=pypi_0
flatbuffers=1.12=pypi_0
fonttools=4.42.1=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
ftfy=6.1.1=pypi_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.10=pypi_0
gitpython=3.1.32=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.32.0=pypi_0
h11=0.14.0=pypi_0
h5py=2.10.0=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.1=pypi_0
huggingface-hub=0.16.4=pypi_0
identify=2.5.27=pypi_0
idna=3.4=py38haa95532_0
imageio=2.31.2=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.1=pypi_0
intel-openmp=2023.1.0=h59b6b97_46319
iopath=0.1.10=pypi_0
ipython=8.12.2=pypi_0
jedi=0.19.0=pypi_0
jinja2=3.1.2=py38haa95532_0
joblib=1.3.2=pypi_0
jpeg=9e=h2bbff1b_1
jsonschema=4.19.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
kaggle=1.5.16=pypi_0
kiwisolver=1.4.5=pypi_0
langcodes=3.3.0=pypi_0
lazy-loader=0.3=pypi_0
lerc=3.0=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.3.129=0
libcurand-dev=10.3.3.129=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
magic-filter=1.0.11=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.6=py38hf11a4ad_1
mkl_random=1.2.2=py38hf11a4ad_1
mpmath=1.3.0=py38haa95532_0
multidict=6.0.4=pypi_0
murmurhash=1.0.9=pypi_0
networkx=3.1=py38haa95532_0
nltk=3.8.1=pypi_0
nodeenv=1.8.0=pypi_0
numpy=1.19.5=pypi_0
omegaconf=2.3.0=pypi_0
openai=0.28.0=pypi_0
opencv-python-headless=4.5.5.64=pypi_0
opendatasets=0.1.22=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.10=h2bbff1b_2
packaging=23.1=pypi_0
pandas=1.3.3=pypi_0
parso=0.8.3=pypi_0
pathy=0.10.2=pypi_0
peft=0.5.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py38hd77b12b_0
pip=23.2.1=py38haa95532_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.10.0=pypi_0
plotly=5.16.1=pypi_0
portalocker=2.7.0=pypi_0
pre-commit=3.3.3=pypi_0
preshed=3.0.8=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=13.0.0=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.3.0=pypi_0
pydantic-core=2.6.3=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.16.1=pypi_0
pympler=1.0.1=pypi_0
pyopenssl=23.2.0=py38haa95532_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38haa95532_0
python=3.8.17=h1aa4202_0
python-dateutil=2.8.2=pypi_0
python-magic=0.4.27=pypi_0
python-slugify=8.0.1=pypi_0
python-telegram-bot=13.13=pypi_0
pytorch=2.0.1=py3.8_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywavelets=1.4.1=pypi_0
pywin32=306=pypi_0
pyyaml=6.0.1=pypi_0
referencing=0.30.2=pypi_0
regex=2023.8.8=pypi_0
requests=2.31.0=py38haa95532_0
rich=13.5.2=pypi_0
rpds-py=0.9.2=pypi_0
safetensors=0.3.3=pypi_0
salesforce-lavis=1.0.1=pypi_0
scikit-image=0.18.3=pypi_0
scikit-learn=1.3.1=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
semver=3.0.1=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.0.0=py38haa95532_0
six=1.15.0=pypi_0
smart-open=6.3.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
spacy=3.6.1=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.4=pypi_0
sqlite=3.41.2=h2bbff1b_0
srsly=2.4.7=pypi_0
stack-data=0.6.2=pypi_0
streamlit=1.14.0=pypi_0
sympy=1.11.1=py38haa95532_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.3=pypi_0
tensorboard=2.4.0=pypi_0
tensorflow-estimator=2.4.0=pypi_0
text-unidecode=1.3=pypi_0
thinc=8.1.12=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.4.12=pypi_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.13.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torchaudio=2.0.2=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.25.0=pypi_0
typer=0.9.0=pypi_0
typing_extensions=4.7.1=py38haa95532_0
tzdata=2023.3=pypi_0
tzlocal=4.3.1=pypi_0
urllib3=1.26.16=py38haa95532_0
validators=0.21.2=pypi_0
vc=14.2=h21ff451_1
virtualenv=20.24.3=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wasabi=1.1.2=pypi_0
watchdog=3.0.0=pypi_0
wcwidth=0.2.6=pypi_0
webdataset=0.2.48=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.4.2=h8cc25b3_0
yarl=1.9.2=pypi_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0

numpy~=1.19.5
pycocoevalcap~=1.2
pycocotools~=2.0.7
torchvision~=0.15.2
pandas~=1.3.3
pyyaml~=6.0.1
iopath~=0.1.10
einops~=0.6.1
timm~=0.4.12
matplotlib~=3.4.3
scipy~=1.10.1
scikit-image~=0.18.3
regex~=2023.8.8
tifffile~=2023.7.10
pillow~=9.4.0
packaging~=23.1
tensorboard~=2.4.0
omegaconf~=2.3.0
transformers~=4.25.0
fairscale~=0.4.4
ftfy~=6.1.1
tqdm~=4.66.1
spacy~=3.6.1
diffusers~=0.16.0
ipython~=8.12.2
webdataset~=0.2.48
decord~=0.6.0
opendatasets~=0.1.22
requests~=2.31.0
zlib~=1.2.13
python-magic~=0.4.27
peft~=0.5.0
aiohttp~=3.8.5
aiogram~=2.14
seaborn~=0.12.2